refactor: split LC013 and LC039 analysis helpers
- split disposed-context-query origin resolution and type checks
- split nested-save-changes boundary and context resolution helpers
- keep analyzer behavior unchanged while reducing local complexity
