+++ src/LinqContraband/Analyzers/ChangeTrackingAndContextLifetime/LC039_NestedSaveChanges/NestedSaveChangesBoundaryAnalysis.cs
@@ -1,0 +1,27 @@
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.Diagnostics;
+
+namespace LinqContraband.Analyzers.LC039_NestedSaveChanges;
+
+public sealed partial class NestedSaveChangesAnalyzer
+{
+    private sealed partial class AnalysisState
+    {
+        private static bool HasTransactionBoundaryBetween(int[] boundaries, int left, int right)
+        {
+            foreach (var boundary in boundaries)
+            {
+                if (boundary > left && boundary < right)
+                    return true;
+            }
+
+            return false;
+        }
+
+        private static void Report(CompilationAnalysisContext context, InvocationRecord record, ISymbol contextSymbol, string methodName)
+        {
+            context.ReportDiagnostic(
+                Diagnostic.Create(Rule, record.Location, contextSymbol.Name, methodName));
+        }
+    }
+}
